add Hero component, Footer component, and KelasProject component; update page layout and dependencies
