+++ app/class-fase-e/page.tsx
@@ -1,3 +1,4 @@
+import { Metadata } from "next";
 import React from "react";
 
 const KenapaPilihPPLG = () => {
@@ -11,12 +12,12 @@ const KenapaPilihPPLG = () => {
         {/* Card untuk Kenapa Memilih PPLG */}
         <div className="relative group">
           <img
-            src="/12.png" // Ganti dengan foto kelas
+            src="/class/10.jpg" // Ganti dengan foto kelas
             alt="Foto Kelas"
             className="w-full h-80 object-cover rounded-lg shadow-lg transform transition duration-500 group-hover:scale-105"
           />
           <a
-            href="https://www.instagram.com/d.sonere/" // Ganti dengan link IG kelas
+            href="https://www.instagram.com/epplg21_/" // Ganti dengan link IG kelas
             target="_blank"
             rel="noopener noreferrer"
             className="absolute inset-0 bg-black bg-opacity-50 flex justify-center items-center opacity-0 group-hover:opacity-100 transition-opacity duration-300"
@@ -29,28 +30,22 @@ const KenapaPilihPPLG = () => {
         <div className="bg-gradient-to-r from-blue-700 to-blue-900 p-8 rounded-xl shadow-lg">
           <h3 className="text-3xl font-bold mb-4">Alasan Memilih PPLG</h3>
           <ul className="list-disc pl-6 space-y-4 text-lg">
-          <li>Karena akuh maw jadi hengker NASA (zahir sang sepuh).</li>
-              <li>pelarian setelah ditolak dkv (jose).</li>
-              <li>Iseng(Si Tamvan kls 12).</li>
-              <li>karna gabut().</li>
-              <li>karna saya sangat antusias dengan teknologi().</li>
-              <li>karna gua ganteng().</li>
-              <li>karna gua pengen().</li>
-              <li>karna gua pengen jadi programer().</li>
-              <li>karna mau dapet duit.</li>
-              <li>karna saya ingin tahu cara membuat game.</li>
-              <li>Karena keterpaksaan dan sekarang menjadi biasa.</li>
-              <li>karena tertarik dengan coding terutama membuat web.</li>
-              <li>aku suka coding bg.</li>
-              <li>Karena ingin masuk sebagai programer.</li>
-              <li>Karena peluang karier di berbagai bidang teknologi, seperti developer software, game, dan web, sangat besar. Selain itu, PPLG memberikan kesempatan untuk meningkatkan wawasan dan keterampilan, mulai dari mempelajari coding, desain, hingga pengembangan teknologi secara mendalam dan terstruktur..</li>
-
+            <li>Karena rasa penasaran akan dunia software, dan ingin mempelajarinya secara mendalam 🤤🤤 (Lidya Pro Gemink pepeelgeh).</li>
+            <li>Karena gajinya gede trus kerja gampang  (Manusiaa kelas 10).</li>
+            <li>pengen jadi hengker(canda) mau jadi ui/ux designer tapi gapaham teknologi jadi nyoba eksplor disini (Syaban kelas 10).</li>
+            <li>Karena PPLG itu seru dan sangat enak untuk belajar nya (Fadhil).</li>
+            <li>bisa mengetahui tentang teknologi secara luas dan inovatif (Magfirawrrri).</li>
+            <li>alasannya karena dulu pernah nyoba yg akhirnya malah tertarik (nata).</li>
+            <li>Karena mau jadi web desain PPLG (Selly meliyanti X PPLG).</li>
+            <li>Mau belajar coding, dan abang saya menyarankan ke saya untuk masuk ke jurusan RPL/PPLG (Muhammad Raafi X PPLG).</li>
+            <li>Karena minat dari smp sudah mengarah ke IT (Huga Ghaisan Syahputra Kelas 10 PP).</li>
+            <li>Karna saya ingin tahu cara membuat game (anomali).</li>
           </ul>
         </div>
 
         {/* Card Cita-Cita Kelas 10 */}
         <div className="bg-gradient-to-r from-green-700 to-green-900 p-8 rounded-xl shadow-lg">
-          <h3 className="text-3xl font-bold mb-4">Cita-Cita dari Kelas 12</h3>
+          <h3 className="text-3xl font-bold mb-4">Cita-Cita dari Kelas 10</h3>
           <ul className="list-disc pl-6 space-y-4 text-lg">
             <li>Menjadi programmer sukses.</li>
             <li>Menjadi ahli di bidang teknologi informasi.</li>
@@ -63,5 +58,12 @@ const KenapaPilihPPLG = () => {
     </div>
   );
 };
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: "Kelas Fase E",
+    description: "Kelas Fase E adalah kelas yang berada di jurusan PPLG.",
+  }
+}
 
 export default KenapaPilihPPLG;
