+++ app/layout.tsx
@@ -1,23 +1,19 @@
 import type { Metadata } from "next";
-import localFont from "next/font/local";
+import { Poppins } from "next/font/google";
 import "./globals.css";
-
-const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
-  variable: "--font-geist-sans",
-  weight: "100 900",
-});
-const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
-  variable: "--font-geist-mono",
-  weight: "100 900",
-});
+import { Navbar } from "@/components/Navbar";
+import Footer from "@/components/Footer";
 
 export const metadata: Metadata = {
   title: "PPLG CLASS",
-  description: "Generated by create next app",
+  description: "PPLG adalah salah satu jurusan di SMK Negeri 21 Jakarta.",
+  authors: [{name: "bolehngopi"}, {name: "rezezi"}],
 };
 
+const poppins = Poppins({
+  weight: "400",
+  subsets: ['latin'],
+})
 export default function RootLayout({
   children,
 }: Readonly<{
@@ -26,9 +22,11 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+        className={`${poppins.className} antialiased`}
       >
+        <Navbar />
         {children}
+        <Footer />
       </body>
     </html>
   );
