+++ data/index.ts
@@ -1,0 +1,179 @@
+export const kelas10Projects = [
+  {
+    name: "Kasir",
+    owner: "Huga Ghaisan Syahputra",
+    description: "Kasir Dekstop",
+    image: "/projects/huga.jpg", // Pastikan file gambar tersedia di direktori public
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Dimas Syahreza",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/dmas.jpg", // Pastikan file gambar tersedia di direktori public
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Portofolio",
+    owner: "Habibi",
+    description: "Portofolio",
+    image: "/projects/bibi.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Muhammad Fathurrahman Wahid",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/fathr.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Cyber",
+    owner: "Magfi Adi Radza Putra",
+    description: "pico ctf",
+    image: "/projects/magfi.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Game Unity",
+    owner: "Muhammad Fathurrahman Wahid",
+    description: "Game sederhana",
+    image: "/projects/tur.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Charissa Chartalia Chiaravita",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/charisa.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Fadhil Danendra Hadie Brata",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/fadhli.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Earlyta Dwi Anggraeni",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/dwi.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  {
+    name: "Sistem Kasir Desktop",
+    owner: "Raisya Kamilah",
+    description: "Aplikasi kasir sederhana berbasis desktop",
+    image: "/projects/kamil.jpg", // Tambahkan gambar
+    link: "", // Link dapat ditambahkan nanti
+  },
+  // ... Proyek lainnya untuk kelas 10
+];
+
+export const kelas11Projects = [
+  {
+    name: "Portofolio",
+    owner: "Ahmad dhani",
+    description: "Website Portofolio",
+    image: "/projects/dan.png",
+    link: "https://www.canva.com/design/DAGYyoYJ3BE/-zxHrNQTjRfiXvvBiFcVaw/edit?utm_content=DAGYyoYJ3BE&utm_campaign=designshare&utm_medium=link2&utm_source=sharebutton", // Link ke Vercel
+  },
+  {
+    name: "Portofolio",
+    owner: "Ahmad Habibi Arouf",
+    description: "Website Portofolio",
+    image: "/projects/bie.png",
+    link: "https://abieport.vercel.app/", // Link ke Vercel
+  },
+  {
+    name: "Portofolio",
+    owner: "Fajar Andhika",
+    description: "Website Portofolio",
+    image: "/projects/fajar.png",
+    link: "https://fan-png.vercel.app/", // Link ke Vercel
+  },
+  {
+    name: "Game",
+    owner: "Antonius i wayan ",
+    description: "Game Platformer",
+    image: "/projects/an.png",
+    link: "https://ekorrisu.itch.io/flag-pairs-game/", // Link ke Vercel
+  },
+  {
+    name: "Portofolio",
+    owner: "Alfat Ghifary",
+    description: "Website Portofolio",
+    image: "/projects/pat.png",
+    link: "https://portofolio-sederhana.netlify.app/", // Link ke Vercel
+  },
+  {
+    name: "Portofolio",
+    owner: "Adrian Adiputra",
+    description: "Website Portofolio",
+    image: "/projects/dri.png",
+    link: "https://myportfolioadrian.vercel.app/?fbclid=PAZXh0bgNhZW0CMTEAAaYyFV6wNi_zWTondsiX3UJYQeOSHW9RFAG3bmGuK-qL-K8b9krHPSqHuio_aem_VnfAA6RkZ_pgXIv40FjUHQ",
+  },
+  {
+    name: "Portofolio",
+    owner: "Garuda putra ",
+    description: "Website Portofolio",
+    image: "/projects/gar.png",
+    link: "https://portfolio-garuda-putera-eko.vercel.app/",
+  },
+  {
+    name: "Portofolio",
+    owner: "Sami Athari Zahid",
+    description: "Website Portofolio",
+    image: "/projects/sam.png",
+    link: "https://portofolio-psi-umber.vercel.app/?fbclid=PAZXh0bgNhZW0CMTEAAablKjeuhK6cCeawC1zWN1FPLA05cgwcSB4VoMIefGqLC70K1RP6In3lqDA_aem_mlz0SDu2TU1Dv1vn9Egj8g", // Link ke Vercel
+  },
+  {
+    name: "game",
+    owner: "Muhammad Al Mahdi",
+    description: "Game platformer",
+    image: "/projects/madi.png",
+    link: "https://almahdi00.itch.io/adventure-knight-in-fairyland", // Link ke Vercel
+  },
+  {
+    name: "Game",
+    owner: "Dwi Wahyu",
+    description: "Game platformer",
+    image: "/projects/wi.png",
+    link: "https://nyxaria-27.itch.io/veil-of-circuits", // Link ke Vercel
+  },
+  {
+    name: "Game",
+    owner: "Pipit wulandari",
+    description: "Game platformer",
+    image: "/projects/pipit.png",
+    link: "https://pipit6.itch.io/kucing", // Link ke Vercel
+  },
+  {
+    name: "Game",
+    owner: "Muhhamad Fabian",
+    description: "Game platformer",
+    image: "/projects/fabian.png",
+    link: "https://gd.games/instant-builds/2bec8ab5-914a-49d5-af7a-824b15908e61", // Link ke Vercel
+  },
+  {
+    name: "Game",
+    owner: "Sami Athari Zahid",
+    description: "Game platformer",
+    image: "/projects/sa.png",
+    link: "https://sami-athari.itch.io/nameless-cat", // Link ke Vercel
+  },
+  // ... Proyek lainnya untuk kelas 11
+];
+
+export const kelas12Projects = [
+  {
+    name: "Website Perpustakaan SMKN21",
+    owner: "Zahir",
+    description: "Website Perpustakaan SMKN21",
+    image: "/projects/zahir.png", // Pastikan file gambar tersedia di direktori public
+    link: "https://perpustakaan-smkn21jakarta.vercel.app/", // Contoh link
+  },
+  // Tambahkan proyek lain di sini jika ada
+];
