+++ package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "next": "^15.0.4",
     "react": "19.0.0-rc-66855b96-20241106",
-    "react-dom": "19.0.0-rc-66855b96-20241106"
+    "react-dom": "19.0.0-rc-66855b96-20241106",
+    "react-icons": "^5.4.0"
   },
   "devDependencies": {
     "@types/node": "^20",
